fix : CI functions moved to pycm_ci.py #237
